move arcade/asset/data to arcade/data so it can be shared between asset and user
