+++ data/item_test.go
@@ -19,8 +19,8 @@ import (
 
 	"arcadium.dev/arcade"
 	"arcadium.dev/arcade/asset"
-	"arcadium.dev/arcade/asset/data"
-	"arcadium.dev/arcade/asset/data/postgres"
+	"arcadium.dev/arcade/data"
+	"arcadium.dev/arcade/data/postgres"
 )
 
 func TestItemsList(t *testing.T) {
